requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlmodel>=0.0.8
sqlalchemy>=2.0.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
pydantic>=1.8.0
supabase>=1.0.3 